Stop tracking a minified version.
Tracking a minified version is complicated,
because you have to keep it in sync with every commit.

In addition, it is not even clear what minifier has been used
for this minification, so when different developers
update it, it might change fully, which is a lot
of noise in the commits.